requirements.txt
anyascii==0.3.2
apturl==0.5.2
asgiref==3.8.1
attrs==23.2.0
bcp47==0.0.4
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brlapi==0.8.3
certifi==2024.2.2
chardet==4.0.0
charset-normalizer==3.3.2
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-modelcluster==6.3
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
draftjs-exporter==5.0.0
drf-spectacular==0.27.2
drf-yasg==1.21.7
duplicity==0.8.21
et-xmlfile==1.1.0
fasteners==0.14.1
filetype==1.2.0
future==0.18.2
httplib2==0.20.2
idna==3.7
importlib-metadata==4.6.4
inflection==0.5.1
jeepney==0.7.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==23.5.0
l18n==2021.3
laces==0.1.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
MarkupSafe==2.0.1
mercurial==6.1.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.1.3
packaging==24.0
paramiko==2.9.3
pexpect==4.8.0
pillow==10.3.0
pillow_heif==0.16.0
protobuf==3.12.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
PyGObject==3.42.1
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==0.19.2
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
pyzbar==0.1.9
referencing==0.35.1
reportlab==3.6.8
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.18.1
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
systemd-python==234
telepath==0.3.1
texttable==1.6.4
transliterate==1.10.2
typing_extensions==4.12.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.19
usb-creator==0.3.7
wadllib==1.3.6
wagtail==6.1.2
wagtail-modeladmin==2.0.0
wagtailvideos==6.1.1
websocket-client==1.2.3
whitenoise==6.5.0
Willow==1.8.0
xdg==5
xkit==0.0.0
zipp==3.16.2
